+++ bevy_asset_loader/examples/atlas_from_grid.rs
@@ -13,8 +13,11 @@ fn main() {
         )
         .add_collection_to_loading_state::<_, MyAssets>(MyStates::AssetLoading)
         .insert_resource(Msaa::Off)
-        .add_system(draw_atlas.in_schedule(OnEnter(MyStates::Next)))
-        .add_system(animate_sprite_system.run_if(in_state(MyStates::Next)))
+        .add_systems(OnEnter(MyStates::Next), draw_atlas)
+        .add_systems(
+            Update,
+            animate_sprite_system.run_if(in_state(MyStates::Next)),
+        )
         .run();
 }