+++ bevy_asset_loader/examples/failure_state.rs
@@ -12,9 +12,9 @@ fn main() {
                 .on_failure_continue_to_state(MyStates::ErrorScreen),
         )
         .add_collection_to_loading_state::<_, MyAssets>(MyStates::AssetLoading)
-        .add_system(timeout.run_if(in_state(MyStates::AssetLoading)))
-        .add_system(fail.in_schedule(OnEnter(MyStates::Next)))
-        .add_system(ok.in_schedule(OnEnter(MyStates::ErrorScreen)))
+        .add_systems(Update, timeout.run_if(in_state(MyStates::AssetLoading)))
+        .add_systems(OnEnter(MyStates::Next), fail)
+        .add_systems(OnEnter(MyStates::ErrorScreen), ok)
         .run();
 }