+++ bevy_asset_loader/examples/no_loading_state.rs
@@ -22,8 +22,8 @@ fn main() {
         // This requires the extension trait `AssetCollectionApp` to be in scope.
         .init_collection::<ImageAssets>()
         // This system listens for mouse clicks and then loads + inserts the AudioAssets collection
-        .add_system(load_and_play_audio)
-        .add_system(draw.on_startup())
+        .add_systems(Startup, draw)
+        .add_systems(Update, load_and_play_audio)
         .run();
 }
 
@@ -37,8 +37,10 @@ fn load_and_play_audio(world: &mut World) {
         world.init_collection::<AudioAssets>();
 
         let audio_assets = world.get_resource::<AudioAssets>().unwrap();
-        let audio = world.get_resource::<Audio>().unwrap();
-        audio.play(audio_assets.background.clone());
+        world.spawn(AudioBundle {
+            source: audio_assets.background.clone(),
+            settings: PlaybackSettings::LOOP,
+        });
     }
 }