+++ bevy_asset_loader/src/loading_state.rs
@@ -1,15 +1,17 @@
 mod dynamic_asset_systems;
 mod systems;
 
-use bevy::app::{App, CoreSet, IntoSystemAppConfig, Plugin};
+use bevy::app::{App, Plugin};
 use bevy::asset::{Asset, HandleUntyped};
-use bevy::ecs::schedule::{
-    common_conditions::in_state, IntoSystemConfig, IntoSystemSetConfig, NextState, OnEnter, States,
-    SystemSet,
+use bevy::ecs::{
+    schedule::{
+        common_conditions::in_state, BoxedScheduleLabel, IntoSystemConfigs, IntoSystemSetConfig,
+        NextState, OnEnter, ScheduleLabel, State, States, SystemSet,
+    },
+    system::Resource,
+    world::FromWorld,
 };
-use bevy::ecs::schedule::{BoxedScheduleLabel, ScheduleLabel, State};
-use bevy::ecs::system::Resource;
-use bevy::ecs::world::FromWorld;
+use bevy::prelude::{StateTransition, Update};
 use bevy::utils::{default, HashMap, HashSet};
 use std::any::TypeId;
 use std::array::IntoIter;
@@ -50,21 +52,23 @@ use crate::loading_state::systems::{apply_internal_state_transition, run_loading
 /// fn main() {
 ///     App::new()
 ///         .add_state::<GameState>()
-///         .add_plugins(MinimalPlugins)
+///         .add_plugins((MinimalPlugins, AssetPlugin::default()))
 /// #       .init_resource::<iyes_progress::ProgressCounter>()
-///         .add_plugin(AssetPlugin::default())
 ///         .add_loading_state(LoadingState::new(GameState::Loading)
 ///             .continue_to_state(GameState::Menu)
 ///         )
 ///         .add_collection_to_loading_state::<_, AudioAssets>(GameState::Loading)
 ///         .add_collection_to_loading_state::<_, ImageAssets>(GameState::Loading)
-///         .add_system(play_audio.in_schedule(OnEnter(GameState::Menu)))
+///         .add_systems(OnEnter(GameState::Menu), play_audio)
 /// #       .set_runner(|mut app| app.update())
 ///         .run();
 /// }
 ///
-/// fn play_audio(audio_assets: Res<AudioAssets>, audio: Res<Audio>) {
-///     audio.play(audio_assets.background.clone());
+/// fn play_audio(mut commands: Commands, audio_assets: Res<AudioAssets>) {
+///     commands.spawn(AudioBundle {
+///         source: audio_assets.background.clone(),
+///         ..default()
+///     });
 /// }
 ///
 /// #[derive(Clone, Eq, PartialEq, Debug, Hash, Default, States)]
@@ -113,9 +117,8 @@ where
     /// # fn main() {
     ///     App::new()
     /// #       .add_state::<GameState>()
-    /// #       .add_plugins(MinimalPlugins)
+    /// #       .add_plugins((MinimalPlugins, AssetPlugin::default()))
     /// #       .init_resource::<iyes_progress::ProgressCounter>()
-    /// #       .add_plugin(AssetPlugin::default())
     ///         .add_loading_state(
     ///           LoadingState::new(GameState::Loading)
     ///             .continue_to_state(GameState::Menu)
@@ -165,9 +168,8 @@ where
     /// # fn main() {
     ///     App::new()
     /// #       .add_state::<GameState>()
-    /// #       .add_plugins(MinimalPlugins)
+    /// #       .add_plugins((MinimalPlugins, AssetPlugin::default()))
     /// #       .init_resource::<iyes_progress::ProgressCounter>()
-    /// #       .add_plugin(AssetPlugin::default())
     ///         .add_loading_state(
     ///           LoadingState::new(GameState::Loading)
     ///             .continue_to_state(GameState::Menu)
@@ -211,9 +213,8 @@ where
     /// # fn main() {
     ///     App::new()
     /// #       .add_state::<GameState>()
-    /// #       .add_plugins(MinimalPlugins)
+    /// #       .add_plugins((MinimalPlugins, AssetPlugin::default()))
     /// #       .init_resource::<iyes_progress::ProgressCounter>()
-    /// #       .add_plugin(AssetPlugin::default())
     ///         .add_loading_state(
     ///           LoadingState::new(GameState::Loading)
     ///             .continue_to_state(GameState::Menu)
@@ -285,9 +286,8 @@ where
     /// # fn main() {
     ///     App::new()
     /// #       .add_state::<GameState>()
-    /// #       .add_plugins(MinimalPlugins)
+    /// #       .add_plugins((MinimalPlugins, AssetPlugin::default()))
     /// #       .init_resource::<iyes_progress::ProgressCounter>()
-    /// #       .add_plugin(AssetPlugin::default())
     ///         .add_loading_state(
     ///           LoadingState::new(GameState::Loading)
     ///             .continue_to_state(GameState::Menu)
@@ -344,13 +344,13 @@ where
         app.init_resource::<DynamicAssetCollections<S>>();
         #[cfg(feature = "standard_dynamic_assets")]
         if !app.is_plugin_added::<RonAssetPlugin<StandardDynamicAssetCollection>>() {
-            app.add_plugin(RonAssetPlugin::<StandardDynamicAssetCollection>::new(
+            app.add_plugins(RonAssetPlugin::<StandardDynamicAssetCollection>::new(
                 &self.standard_dynamic_asset_collection_file_endings,
             ));
         }
 
         if !app.is_plugin_added::<InternalAssetLoaderPlugin<S>>() {
-            app.add_plugin(InternalAssetLoaderPlugin::<S>::new());
+            app.add_plugins(InternalAssetLoaderPlugin::<S>::new());
         }
 
         app.init_resource::<LoadingStateSchedules<S>>();
@@ -372,26 +372,21 @@ where
             ));
 
         if configure_loading_state {
-            app.add_systems((
-                resume_to_loading_asset_collections::<S>
-                    .in_schedule(loading_state_schedule.clone())
-                    .in_set(InternalLoadingStateSet::ResumeDynamicAssetCollections),
-                initialize_loading_state::<S>
-                    .in_schedule(loading_state_schedule.clone())
-                    .in_set(InternalLoadingStateSet::Initialize),
-                resume_to_finalize::<S>
-                    .in_schedule(loading_state_schedule.clone())
-                    .in_set(InternalLoadingStateSet::CheckAssets),
-                finish_loading_state::<S>
-                    .in_schedule(loading_state_schedule.clone())
-                    .in_set(InternalLoadingStateSet::Finalize),
-            ))
-            .add_system(reset_loading_state::<S>.in_schedule(OnEnter(self.loading_state.clone())))
-            .configure_set(
-                LoadingStateSet(self.loading_state.clone())
-                    .after(CoreSet::StateTransitions)
-                    .before(CoreSet::Update),
-            );
+            app.add_systems(
+                loading_state_schedule.clone(),
+                (
+                    resume_to_loading_asset_collections::<S>
+                        .in_set(InternalLoadingStateSet::ResumeDynamicAssetCollections),
+                    initialize_loading_state::<S>.in_set(InternalLoadingStateSet::Initialize),
+                    resume_to_finalize::<S>.in_set(InternalLoadingStateSet::CheckAssets),
+                    finish_loading_state::<S>.in_set(InternalLoadingStateSet::Finalize),
+                ),
+            )
+            .add_systems(
+                OnEnter(self.loading_state.clone()),
+                reset_loading_state::<S>,
+            )
+            .configure_set(Update, LoadingStateSet(self.loading_state.clone()));
             let mut loading_state_schedule = app.get_schedule_mut(loading_state_schedule).unwrap();
             loading_state_schedule
                 .configure_set(
@@ -419,17 +414,24 @@ where
                         .run_if(in_state(InternalLoadingState::<S>::Finalize)),
                 );
 
+            #[cfg(feature = "standard_dynamic_assets")]
+            app.register_dynamic_asset_collection::<_, StandardDynamicAssetCollection>(
+                self.loading_state.clone(),
+            );
+
             #[cfg(feature = "progress_tracking")]
-            app.add_system(
+            app.add_systems(
+                Update,
                 run_loading_state::<S>
                     .in_set(TrackedProgressSet)
-                    .in_base_set(LoadingStateSet(self.loading_state.clone()))
+                    .in_set(LoadingStateSet(self.loading_state.clone()))
                     .run_if(in_state(self.loading_state)),
             );
             #[cfg(not(feature = "progress_tracking"))]
-            app.add_system(
+            app.add_systems(
+                Update,
                 run_loading_state::<S>
-                    .in_base_set(LoadingStateSet(self.loading_state.clone()))
+                    .in_set(LoadingStateSet(self.loading_state.clone()))
                     .run_if(in_state(self.loading_state)),
             );
         }
@@ -442,10 +444,8 @@ where
     }
 }
 
-/// This set runs after [`CoreSet::StateTransitions`] and before [`CoreSet::Update`].
-/// Systems in this set check the loading state of assets and will change the [`InternalLoadingState`] accordingly.
+///  Systems in this set check the loading state of assets and will change the [`InternalLoadingState`] accordingly.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, SystemSet)]
-#[system_set(base)]
 pub(crate) struct LoadingStateSet<S: States>(S);
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, SystemSet)]
@@ -572,9 +572,8 @@ pub trait LoadingStateAppExt {
     /// # fn main() {
     ///     App::new()
     /// #       .add_state::<GameState>()
-    /// #       .add_plugins(MinimalPlugins)
+    /// #       .add_plugins((MinimalPlugins, AssetPlugin::default()))
     /// #       .init_resource::<iyes_progress::ProgressCounter>()
-    /// #       .add_plugin(AssetPlugin::default())
     ///         .add_loading_state(
     ///           LoadingState::new(GameState::Loading)
     ///             .continue_to_state(GameState::Menu)
@@ -608,6 +607,15 @@ pub trait LoadingStateAppExt {
         loading_state: S,
     ) -> &mut Self;
 
+    /// Register a new [`DynamicAssetCollection`] to be handled in the loading state
+    ///
+    /// You do not need to call this for [`StandardDynamicAssetCollection`], only if you want to use
+    /// your own dynamic asset collection types.
+    fn register_dynamic_asset_collection<S: States, C: DynamicAssetCollection + Asset>(
+        &mut self,
+        loading_state: S,
+    ) -> &mut Self;
+
     /// Register files to be loaded as a certain type of [`DynamicAssetCollection`]
     ///
     /// During the loading state, the given dynamic asset collections will be loaded and their
@@ -629,10 +637,9 @@ pub trait LoadingStateAppExt {
     /// # use bevy::asset::AssetPlugin;
     /// # fn main() {
     ///     App::new()
-    /// #       .add_plugins(MinimalPlugins)
+    /// #       .add_plugins((MinimalPlugins, AssetPlugin::default()))
     /// #       .add_state::<GameState>()
     /// #       .init_resource::<iyes_progress::ProgressCounter>()
-    /// #       .add_plugin(AssetPlugin::default())
     ///         .add_loading_state(
     ///           LoadingState::new(GameState::Loading)
     ///             .continue_to_state(GameState::Menu)
@@ -685,16 +692,31 @@ impl LoadingStateAppExt for App {
         &mut self,
         loading_state: S,
     ) -> &mut Self {
-        self.add_system(
-            start_loading_collection::<S, A>.in_schedule(OnEnterInternalLoadingState(
-                loading_state.clone(),
-                InternalLoadingState::LoadingAssets,
-            )),
+        self.add_systems(
+            OnEnterInternalLoadingState(loading_state.clone(), InternalLoadingState::LoadingAssets),
+            start_loading_collection::<S, A>,
         )
-        .add_system(
-            check_loading_collection::<S, A>
-                .in_schedule(LoadingStateSchedule(loading_state))
-                .in_set(InternalLoadingStateSet::CheckAssets),
+        .add_systems(
+            LoadingStateSchedule(loading_state),
+            check_loading_collection::<S, A>.in_set(InternalLoadingStateSet::CheckAssets),
+        )
+    }
+
+    fn register_dynamic_asset_collection<S: States, C: DynamicAssetCollection + Asset>(
+        &mut self,
+        loading_state: S,
+    ) -> &mut Self {
+        self.add_systems(
+            OnEnterInternalLoadingState(
+                loading_state.clone(),
+                InternalLoadingState::LoadingDynamicAssetCollections,
+            ),
+            load_dynamic_asset_collections::<S, C>,
+        )
+        .add_systems(
+            LoadingStateSchedule(loading_state),
+            check_dynamic_asset_collections::<S, C>
+                .in_set(InternalLoadingStateSet::CheckDynamicAssetCollections),
         )
     }
 
@@ -708,20 +730,7 @@ impl LoadingStateAppExt for App {
             .get_resource_mut::<DynamicAssetCollections<S>>()
             .unwrap();
 
-        if dynamic_asset_collections.register_file::<C>(loading_state.clone(), file) {
-            self.add_system(load_dynamic_asset_collections::<S, C>.in_schedule(
-                OnEnterInternalLoadingState(
-                    loading_state.clone(),
-                    InternalLoadingState::LoadingDynamicAssetCollections,
-                ),
-            ))
-            .add_system(
-                check_dynamic_asset_collections::<S, C>
-                    .in_schedule(LoadingStateSchedule(loading_state))
-                    .in_set(InternalLoadingStateSet::CheckDynamicAssetCollections),
-            );
-        }
-
+        dynamic_asset_collections.register_file::<C>(loading_state.clone(), file);
         self
     }
 
@@ -729,10 +738,10 @@ impl LoadingStateAppExt for App {
         &mut self,
         loading_state: S,
     ) -> &mut Self {
-        self.add_system(init_resource::<A>.in_schedule(OnEnterInternalLoadingState(
-            loading_state,
-            InternalLoadingState::Finalize,
-        )))
+        self.add_systems(
+            OnEnterInternalLoadingState(loading_state, InternalLoadingState::Finalize),
+            init_resource::<A>,
+        )
     }
 }
 
@@ -756,6 +765,6 @@ where
     S: States,
 {
     fn build(&self, app: &mut App) {
-        app.add_system(apply_internal_state_transition::<S>.in_base_set(CoreSet::StateTransitions));
+        app.add_systems(StateTransition, apply_internal_state_transition::<S>);
     }
 }