+++ bevy_asset_loader/tests/multiple_states.rs
@@ -13,21 +13,24 @@ use bevy_asset_loader::prelude::{AssetCollection, LoadingState, LoadingStateAppE
 #[test]
 fn main() {
     App::new()
-        .add_plugins(MinimalPlugins)
-        .add_plugin(AssetPlugin::default())
-        .add_plugin(AudioPlugin::default())
+        .add_plugins((
+            MinimalPlugins,
+            AssetPlugin::default(),
+            AudioPlugin::default(),
+        ))
         .add_state::<Loading>()
         .add_state::<Game>()
         .add_loading_state(LoadingState::new(Game::Booting).continue_to_state(Game::Loading))
         .add_collection_to_loading_state::<_, GameStateCollection>(Game::Booting)
         .add_loading_state(LoadingState::new(Loading::Loading).continue_to_state(Loading::Finalize))
         .add_collection_to_loading_state::<_, LoadingStateCollection>(Loading::Loading)
-        .add_system(go_to_loading_loading.in_schedule(OnEnter(Game::Loading)))
-        .add_system(probe_game_state.in_schedule(OnEnter(Game::Loading)))
-        .add_system(go_to_game_play_loading_done.in_schedule(OnEnter(Loading::Finalize)))
-        .add_system(probe_loading_state.in_schedule(OnEnter(Game::Play)))
-        .add_system(quit.in_set(OnUpdate(Game::Play)))
-        .add_system(timeout)
+        .add_systems(Update, (quit.run_if(in_state(Game::Play)), timeout))
+        .add_systems(
+            OnEnter(Game::Loading),
+            (go_to_loading_loading, probe_game_state),
+        )
+        .add_systems(OnEnter(Loading::Finalize), go_to_game_play_loading_done)
+        .add_systems(OnEnter(Game::Play), probe_loading_state)
         .run();
 }