+++ bevy_asset_loader_derive/src/lib.rs
@@ -175,7 +175,7 @@ fn impl_asset_collection(
     let mut prepare_from_world = quote! {};
     prepare_from_world.append_all(from_world_fields.iter().fold(
         quote!(),
-        |es, _| quote! (#es ::bevy::ecs::world::FromWorld::from_world(world),),
+        |es, ident| quote_spanned! {ident.span() => #es ::bevy::ecs::world::FromWorld::from_world(world),},
     ));
 
     let mut asset_creation = assets.iter().fold(quote!(), |token_stream, asset| {